+++ rules/src/tokens.rs
@@ -117,7 +117,7 @@ pub enum Token {
     Weekday(Weekday),
     When(When),
     AmPm(AmPm),
-    Number(usize),
+    Number(u32),
     IntWord(IntWord),
     TimeInterval(TimeInterval),
     TimeOfDay(TimeOfDay),
